requirements.txt
Django==2.1.5
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
Pillow==5.2.0
pytz==2018.9
